+++ packages/fluentui/react/src/themes/teams/components/Accordion/accordionTitleStyles.ts
@@ -8,8 +8,20 @@ const accordionTitleStyles: ComponentSlotStylesPrepared<AccordionTitleProps> = {
     padding: '.5rem 0',
     cursor: p.disabled ? 'default' : 'pointer'
   }),
+  contentWrapper: ({ props: p }) => ({
+    display: 'grid',
+    gridTemplateColumns: `auto ${p.content ? '1fr' : ''}`
+  }),
   indicator: () => ({
-    userSelect: 'none'
+    userSelect: 'none',
+    alignItems: 'center',
+    display: 'inline-flex',
+    '-ms-grid-column': '1'
+  }),
+  content: () => ({
+    alignItems: 'center',
+    display: 'grid',
+    '-ms-grid-column': '2'
   })
 };